backend/pyproject.toml
[tool.poetry]
name = "jurisagenda-backend"
version = "1.0.0"
description = "JurisAgenda - Sistema Web de Agenda Jurídica (Backend)"
authors = ["Gerência de Arquitetura <dev@jurisagenda.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
channels = "^4.1"
channels-redis = "^4.2"
django-cors-headers = "^4.3"
django-filter = "^23.5"
django-environ = "^0.11"
psycopg2-binary = "^2.9"
boto3 = "^1.34"
celery = "^5.3"
django-celery-beat = "^2.6"
redis = "^5.0"
Pillow = "^10.3"
python-decouple = "^3.8"
drf-spectacular = "^0.27"
argon2-cffi = "^23.1"
gunicorn = "^21.2"
uvicorn = {extras = ["standard"], version = "^0.27"}
python-json-logger = "^2.0"
django-ratelimit = "^4.1"
django-extensions = "^3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1"
pytest-django = "^4.8"
pytest-cov = "^5.0"
factory-boy = "^3.3"
black = "^24.3"
ruff = "^0.3"
mypy = "^1.9"
django-stubs = "^5.0"
ipython = "^8.23"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
strict = false
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "core.settings.local"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings.local"
python_files = ["tests.py", "test_*.py", "*_test.py"]
addopts = "--cov=apps --cov-report=html --cov-fail-under=80"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
